Rename: 포인트 관련 페이지 폴더 구조 수정
