+++ src/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './group'
 export * from './group-management'
+export * from './point'
