+++ src/pages/point/cancel/index.ts
@@ -1,0 +1,1 @@
+export { default as PointCancelModal } from './PointCancelModal'
