+++ src/pages/point/default/components/index.ts
@@ -1,0 +1,3 @@
+export * from './my-point'
+export * from './purchase-point'
+export * from './point-log'
