+++ src/pages/point/default/components/my-point/index.ts
@@ -1,0 +1,1 @@
+export { MyPoint } from './MyPoint'
