+++ src/pages/point/default/components/point-log/index.ts
@@ -1,0 +1,1 @@
+export { PointLog } from './PointLog'
