+++ src/pages/point/default/components/point-log/tab-card/index.ts
@@ -1,0 +1,1 @@
+export { PointLogTabCard } from './TabCard'
