+++ src/pages/point/default/components/purchase-point/index.ts
@@ -1,0 +1,1 @@
+export { PurchasePoint } from './PurchasePoint'
