+++ src/pages/point/default/index.ts
@@ -1,0 +1,1 @@
+export { default as PointPage } from './PointPage'
