+++ src/pages/point/failure/index.ts
@@ -1,0 +1,1 @@
+export { default as PointFailureModal } from './PointFailureModal'
