+++ src/pages/point/index.ts
@@ -1,0 +1,4 @@
+export * from './default'
+export * from './redirect'
+export * from './failure'
+export * from './cancel'
