+++ src/pages/point/redirect/components/index.ts
@@ -1,0 +1,1 @@
+export * from './redirect'
