+++ src/pages/point/redirect/components/redirect/index.ts
@@ -1,0 +1,1 @@
+export { PointRedirect } from './Redirect'
