+++ src/pages/point/redirect/index.ts
@@ -1,0 +1,1 @@
+export { default as PointRedirectPage } from './PointRedirectPage'
